Adds EventsResource get by id method tsk6
Adds a get method that return an event with the provided identifier. It
returns a 200 OK with an event that has the provided identifier. If the
identifier does not corresponds with any event, a 400 Bad Request
response with an error message will be returned. If an unexpected error
happens, a 500 Internal Server Error response will be returned.